refactor(infinite list): move infintie directive in list module
