--- src/framework/theme/components/list/infinite-list/infinite-list.module.ts
@@ -1,14 +1,0 @@
-import { NgModule } from '@angular/core';
-import { NbListModule } from '../list.module';
-import { NbInfiniteListDirective } from './infinite-list.directive';
-
-const componets = [
-  NbInfiniteListDirective,
-];
-
-@NgModule({
-  imports: [ NbListModule ],
-  declarations: componets,
-  exports: componets,
-})
-export class NbInifiniteListModule {}
