--- src/framework/theme/components/list/infinite-list/infintie-list.component.scss
@@ -1,12 +1,0 @@
-:host {
-  display: flex;
-  flex-direction: column;
-}
-
-:host(.element-scroll) {
-  height: 100%;
-
-  nb-list {
-    overflow-y: scroll;
-  }
-}
