+++ src/framework/theme/components/list/list.module.ts
@@ -1,12 +1,14 @@
 import { NgModule } from '@angular/core';
 import { NbListComponent, NbListItemComponent, NbListItemsGroupComponent } from './list.component';
 import { NbListPagerDirective } from './list-pager.directive';
+import { NbInfiniteListDirective } from './infinite-list.directive';
 
 const components = [
   NbListComponent,
   NbListItemComponent,
   NbListItemsGroupComponent,
   NbListPagerDirective,
+  NbInfiniteListDirective,
 ];
 
 @NgModule({
